+++ CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
+
+## 1.3.3 (2024-06-25)
+
+## What's Changed
+
+- Provide a better data product example
+- Fix: eventResources and dataProduct annotation
+
+**Full Changelog**: https://github.com/cap-js/ord/compare/v1.3.2...v1.3.3
+
+## 1.3.2 (2024-06-18)
+
+## What's Changed
+
+- Fix: set cds version to 8
+
+**Full Changelog**: https://github.com/cap-js/ord/compare/v1.3.1...v1.3.2
+
+## 1.3.1 (2024-06-13)
+
+### What's Changed
+
+- Fix: stable dev environment and remove from calesi setup
+- Feat: use relative path for resource definitions URL and compatible with windows
+- Fix: no event or apiResources when build ord files
+
+**Full Changelog**: https://github.com/cap-js/ord/compare/v1.3.0...v1.3.1
+
 ## 1.3.0 (2024-05-20)
 
 ### What's Changed